Adds database to server and proof of work
Client requests targets bits of proof of work from
server, then attaches it to the proposal in a 10 byte
nonce field at the end of the proposal(s).
The proof of work is truncated sha512 (32 bytes) of the
whole encrypted proposal including the nonce.
The database stores proposals as (pubkey, proposal) in
a single table `proposals`.